remove obsolete composer styling code
